Integrate L1 -> L2 fund transfer with vault contracts

+++ smart-contracts/contracts/library/RLPReader.sol
@@ -92,7 +92,7 @@ library RLPReader {
         // 1 byte for the length prefix
         require(item.len == 21, "RLPReader: INVALID_ADDRESS_LENGTH");
 
-        return address(toUint(item));
+        return address(uint160(toUint(item)));
     }
 
     function toUint(RLPItem memory item) internal pure returns (uint256) {
